+++ clients/derivativestradingusdsfutures/tests/unit/websocketapi/api_trade_test.go
@@ -356,7 +356,7 @@ func Test_binancederivativestradingusdsfutureswebsocketapi_TradeAPIService(t *te
 
 		conn.Listen()
 
-		responseChan, errorChan, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float32(1.0)).Price(float32(1.0)).ExecuteAsync()
+		responseChan, errorChan, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float64(1.0)).Price(float64(1.0)).ExecuteAsync()
 		require.NoError(t, err)
 
 		<-mockWS.HasSentChan
@@ -411,7 +411,7 @@ func Test_binancederivativestradingusdsfutureswebsocketapi_TradeAPIService(t *te
 
 		resultChan := make(chan common.ResultWebsocket[models.ModifyOrderResponse], 1)
 		go func() {
-			resp, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float32(1.0)).Price(float32(1.0)).Execute()
+			resp, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float64(1.0)).Price(float64(1.0)).Execute()
 			resultChan <- common.ResultWebsocket[models.ModifyOrderResponse]{Value: resp, Err: err}
 		}()
 
@@ -550,7 +550,7 @@ func Test_binancederivativestradingusdsfutureswebsocketapi_TradeAPIService(t *te
 		done := make(chan struct{})
 
 		go func() {
-			respChan, _, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float32(1.0)).Price(float32(1.0)).ExecuteAsync()
+			respChan, _, err := mockClient.WebsocketAPI.TradeAPI.ModifyOrder().Symbol("BTCUSDT").Side(models.ModifyOrderSideParameterBuy).Quantity(float64(1.0)).Price(float64(1.0)).ExecuteAsync()
 			if err != nil {
 				var wsErr *common.WebSocketError
 				if errors.As(err, &wsErr) {